+++ assignment1/sharedRegions/ContestantsBench.java
@@ -6,12 +6,12 @@ import assignment1.entities.*;
 public class ContestantsBench {
     private GeneralRepos repos;
     private boolean matchOver;
-    private byte callTrial;
-    private byte[][] called;
+    private int callTrial;
+    private int[][] called;
 
     public ContestantsBench(GeneralRepos repos) {
         this.repos = repos;
-        called = new byte[2][SimulPar.NC];
+        called = new int[2][SimulPar.NC];
     }
 
     //Referee
@@ -23,7 +23,7 @@ public class ContestantsBench {
         repos.setRefereeState(RefereeStates.TEAMS_READY);
     }
 
-    public synchronized void declareMatchWinner(short score1, short score2) {
+    public synchronized void declareMatchWinner(int score1, int score2) {
         matchOver = true;
         notifyAll();
         repos.endMatch(score1, score2);
@@ -33,7 +33,7 @@ public class ContestantsBench {
 
     //Coach
 
-    public synchronized byte wait_for_referee_command(short team) {
+    public synchronized int wait_for_referee_command(int team) {
         ((Coach)Thread.currentThread()).setCoachState(CoachStates.WAIT_FOR_REFEREE_COMMAND);
         repos.setCoachState(team, CoachStates.WAIT_FOR_REFEREE_COMMAND);
         while (!matchOver && callTrial==0) {
@@ -45,15 +45,15 @@ public class ContestantsBench {
         return 1;
     }
 
-    public synchronized void callContestants(short team, short[] roster) {
+    public synchronized void callContestants(int team, int[] roster) {
         for (int i = 0; i < SimulPar.NC; i++) called[team][i] = 1;
-        for (short number: roster) called[team][number] = 2;
+        for (int number: roster) called[team][number] = 2;
         notifyAll();
     }
 
     //Contestants
 
-    public synchronized byte seat_at_the_bench(short team, short number) {
+    public synchronized int seat_at_the_bench(int team, int number) {
         Contestant c = (Contestant)Thread.currentThread();
         c.setContestantState(ContestantStates.SEAT_AT_THE_BENCH);
         repos.setContestantState(team, number, ContestantStates.SEAT_AT_THE_BENCH);
@@ -62,7 +62,7 @@ public class ContestantsBench {
             catch (InterruptedException e) {}
         }
         if (matchOver) return 0;
-        byte orders = called[team][number];
+        int orders = called[team][number];
         if (orders==1) {
             c.increaseStrength();
             repos.setContestantStrength(team, number, c.getStrength());
@@ -71,7 +71,7 @@ public class ContestantsBench {
         return orders;
     }
 
-    public synchronized void seatDown(short team, short number) {
+    public synchronized void seatDown(int team, int number) {
         Contestant c = (Contestant)Thread.currentThread();
         c.reduceStrength();
         repos.removeContestant(team, number);
